refactor(product): keep onboarding in existing owners
